Add HQL query

+++ src/main/kotlin/bj/rawsteel/domain/User.kt
@@ -7,6 +7,6 @@ import javax.persistence.Id
  * Created by BaiJiFeiLong@gmail.com at 2018/7/3 下午3:43
  */
 @Entity
-data class User(@Id var id: Int?) {
+data class User(@Id var id: Long?) {
     constructor() : this(null)
 }
